Capitalize Letters like Names SQL Function
This function for SQL Server converts the first letter of each word in Upper Case while the rest of the word is transformed in Lower Casing.

It was written as a final solution for whatever input text, but always you can find better solutions. This code has a commented section for testing the function if you require it.
I wrote it because I couldn't find a simple and true solution for what a I needed.

I hope it will helpful for someone.